refactor: move the Deepnote constants into their own module
DEEPNOTE_TOOLKIT_VERSION, DEEPNOTE_TOOLKIT_PACKAGES, DEEPNOTE_DEFAULT_PORT and
DEEPNOTE_NOTEBOOK_TYPE sat at the bottom of a 400-line types module that imports
`vscode`, so anything wanting a constant had to take the whole tree with it. They
are now a leaf module with no imports, and the fifteen files that used them point
at it.

The E2E venv helper reads the toolkit spec out of the source text — it compiles
under its own tsconfig whose rootDir is test/e2e, so a file from `src` cannot
enter that program (TS6059) — and it now reads four lines instead of scanning
types.ts.

The workflow's pip and venv cache keys hashed types.ts to notice a toolkit
version bump; they hash constants.ts now, or a bump would silently reuse the
cached venv.
