+++ src/kernels/deepnote/constants.ts
@@ -1,0 +1,4 @@
+export const DEEPNOTE_DEFAULT_PORT = 8888;
+export const DEEPNOTE_NOTEBOOK_TYPE = 'deepnote';
+export const DEEPNOTE_TOOLKIT_PACKAGES = ['ipykernel', 'python-lsp-server[all]', 'deepnote-cli'];
+export const DEEPNOTE_TOOLKIT_VERSION = '2.1.1';
