+++ src/kernels/deepnote/deepnoteToolkitInstaller.node.ts
@@ -18,12 +18,8 @@ import {
 } from '../../platform/errors/deepnoteKernelErrors';
 import { logger } from '../../platform/logging';
 import { PythonEnvironment } from '../../platform/pythonEnvironments/info';
-import {
-    DEEPNOTE_TOOLKIT_PACKAGES,
-    DEEPNOTE_TOOLKIT_VERSION,
-    IDeepnoteToolkitInstaller,
-    VenvAndToolkitInstallation
-} from './types';
+import { DEEPNOTE_TOOLKIT_PACKAGES, DEEPNOTE_TOOLKIT_VERSION } from './constants';
+import { IDeepnoteToolkitInstaller, VenvAndToolkitInstallation } from './types';
 
 /**
  * Handles installation of the deepnote-toolkit Python package.
