+++ src/kernels/deepnote/types.ts
@@ -412,9 +412,3 @@ export interface IDeepnoteLspClientManager {
      */
     stopAllClients(token?: vscode.CancellationToken): Promise<void>;
 }
-
-export const DEEPNOTE_TOOLKIT_VERSION = '2.1.1';
-
-export const DEEPNOTE_TOOLKIT_PACKAGES = ['ipykernel', 'python-lsp-server[all]', 'deepnote-cli'];
-export const DEEPNOTE_DEFAULT_PORT = 8888;
-export const DEEPNOTE_NOTEBOOK_TYPE = 'deepnote';
