+++ src/notebooks/deepnote/deepnoteMultiNotebookSplitter.ts
@@ -5,7 +5,7 @@ import { isSingleNotebookDeepnoteFile, splitByNotebooks } from '@deepnote/conver
 import { ITelemetryService } from '../../platform/analytics/types';
 import { ILogger } from '../../platform/logging/types';
 import type { IDeepnoteNotebookEnvironmentMapper } from '../../kernels/deepnote/types';
-import { DEEPNOTE_NOTEBOOK_TYPE } from '../../kernels/deepnote/types';
+import { DEEPNOTE_NOTEBOOK_TYPE } from '../../kernels/deepnote/constants';
 import { readDeepnoteProjectFile } from '../../platform/deepnote/deepnoteProjectFileReader';
 import { allocateSiblingUri } from './deepnoteSiblingFileAllocator';
 import { getFileStem } from './deepnoteNotebookFileFactory';
